Remove useless useEffects and rename popups
