+++ src/components/Overview/Popups/index.ts
@@ -1,3 +1,3 @@
 export * from './EnterValuePopup';
 export * from './EditPopups';
-export * from './PopupForms';
+export * from './FormPopups';
